+++ src/app/modules/secured/secured-app/components/drawer/drawer.component.scss
