+++ src/app/modules/secured/secured-app/components/navbar/navbar.component.ts
