+++ src/app/modules/secured/secured-app/components/sidebar/sidebar.component.scss
