+++ src/app/modules/secured/secured-app/containers/home/components/dropzone/dropzone.component.scss
