+++ src/app/modules/secured/secured-app/containers/home/home.module.ts
@@ -2,8 +2,8 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { HomePageComponent } from './containers/home-page/home-page.component';
 import { HomeRoutingModule } from './home-routing.module';
-import { DropzoneComponent } from '../dropzone/dropzone.component';
-import { ImportModalComponent } from '../import-modal/import-modal.component';
+import { DropzoneComponent } from './components/dropzone/dropzone.component';
+import { ImportModalComponent } from './components/import-modal/import-modal.component';
 import { MatDialogModule } from '@angular/material/dialog';
 import { MatButtonModule } from '@angular/material/button';
