+++ src/app/modules/secured/secured-app/containers/home/models/dropzone.model.ts
