+++ src/app/modules/secured/secured-app/containers/index.ts
@@ -1,5 +1,5 @@
-export * from './navbar/navbar.component'
-export * from './sidebar/sidebar.component'
-export * from './drawer/drawer.component'
-export * from './dropzone/dropzone.component'
-export * from './import-modal/import-modal.component'
+export * from '../components/navbar/navbar.component'
+export * from '../components/sidebar/sidebar.component'
+export * from '../components/drawer/drawer.component'
+export * from './home/components/dropzone/dropzone.component'
+export * from './home/components/import-modal/import-modal.component'
