+++ src/app/modules/secured/secured-app/secured-app-module.ts
@@ -2,8 +2,8 @@ import { CommonModule } from "@angular/common";
 import { NgModule, Optional, SkipSelf } from "@angular/core";
 import { SecuredAppComponent } from "./containers/secured-app-page/secured-app.component";
 import { SecuredAppRoutingModule } from "./secured-app-routing-module";
-import { NavbarComponent } from "./containers/navbar/navbar.component";
-import { SidebarComponent } from "./containers/sidebar/sidebar.component";
+import { NavbarComponent } from "./components/navbar/navbar.component";
+import { SidebarComponent } from "./components/sidebar/sidebar.component";
 import { DrawerComponent } from "./containers";
 
 @NgModule({
